Revert "debug установки валюты - пробуем перевести в u32"
This reverts commit 559424131e49e5a5b11993f7c8a4cba0506f303a.

+++ src/reply.rs
@@ -12,7 +12,7 @@ pub enum Reply {
         description: String,
         payload: String,
         currency: String,
-        amount: f64,
+        amount: u64,
     },
 }
 #[derive(Debug)]
